+++ salesforce/api/tooling.py
@@ -11,6 +11,7 @@ from xml.sax.saxutils import unescape
 
 from ... import requests, util, context
 from ..login import soap_login
+from ..lib.panel import Printer
 
 class ToolingApi():
     def __init__(self, settings, **kwargs):
@@ -815,7 +816,7 @@ class ToolingApi():
             self.settings.get("workbook_field_describe_columns"))+"/"+sobject+".csv"
         print (sobject + " workbook outputdir: " + outputdir)
 
-    def save_component(self, panel, component_attribute, body, is_check_only):
+    def save_component(self, component_attribute, body, is_check_only):
         """ This method contains 5 steps:
             1. Post classid to get MetadataContainerId
             2. Post Component Member
@@ -834,13 +835,11 @@ class ToolingApi():
         """
 
         # Firstly Login
-        util.append_message(panel, "Start login...")
         result = self.login()
         if not result["success"]:
-            util.append_message(panel, "Login failed")
+            Printer.get('log').write("Login failed")
             self.result = result
             return self.result
-        util.append_message(panel, "Login succeed")
 
         # Component Attribute
         component_type = component_attribute["type"]
@@ -848,7 +847,7 @@ class ToolingApi():
         component_body = component_attribute["body"]
 
         if self.settings["check_save_conflict"]:
-            util.append_message(panel, "Start to check saving conflict")
+            Printer.get('log').write("Start to check saving conflict")
             query = "SELECT Id, LastModifiedById, LastModifiedDate " +\
                     "FROM %s WHERE Id = '%s'" % (component_type, component_id)
             result = self.query(query, True)
@@ -876,17 +875,17 @@ class ToolingApi():
 
                 message = "Modified by %s at %s, continue?" % (last_modified_name, lmdate_str)
                 if not sublime.ok_cancel_dialog(message, "Ignore Conflict?"):
-                    util.append_message(panel, "Has conflict, comparing with server...")
+                    Printer.get('log').write("Has conflict, comparing with server...")
                     self.result = {
                         "Operation": "cancel",
                         "Message": "Save operation is cancelled by you due to the conflict"
                     }
                     return self.result
             else:
-                util.append_message(panel, "No conflict, last modified by you at %s" % lmdate_str)
+                Printer.get('log').write("No conflict, last modified by you at %s" % lmdate_str)
 
         # Get MetadataContainerId
-        util.append_message(panel, "Start to fetch MetadataContainerId")
+        Printer.get('log').write("Start to fetch MetadataContainerId")
         data = {  
             "name": "Save" + component_type[4 : len(component_type)] + component_id
         }
@@ -899,7 +898,7 @@ class ToolingApi():
         else:
             # If DUPLICATE Container Id, just delete it and restart this function
             if result["errorCode"] == "DUPLICATE_VALUE":
-                util.append_message(panel, "Start to delete the duplicate MetadataContainerId")
+                Printer.get('log').write("Start to delete the duplicate MetadataContainerId")
                 error_message = result["message"]
                 container_id = error_message[error_message.rindex("1dc"): len(error_message)]
                 delete_result = self.delete(container_url + "/" + container_id)
@@ -911,7 +910,7 @@ class ToolingApi():
                 
                 # We can't reuse the container_id which caused error
                 # Post Request to get MetadataContainerId
-                return self.save_component(panel, component_attribute, body, is_check_only)
+                return self.save_component(component_attribute, body, is_check_only)
 
         # Post ApexComponentMember
         data = {
@@ -923,7 +922,7 @@ class ToolingApi():
         member_result = self.post(url, data)
 
         # Post ContainerAsyncRequest
-        util.append_message(panel, "Start to post ContainerAsyncRequest Request")
+        Printer.get('log').write("Start to post ContainerAsyncRequest Request")
         data = {
             "MetadataContainerId": container_id,
             "IsCheckOnly": is_check_only,
@@ -934,10 +933,9 @@ class ToolingApi():
         request_id = result.get("id")
 
         # Get ContainerAsyncRequest Result
-        util.append_message(panel, "Start to get ContainerAsyncRequest result")
+        Printer.get('log').write("Start to get ContainerAsyncRequest result")
         result = self.get(sync_request_url + "/" + request_id)
         state = result["State"]
-        # print ("Get ContainerAsyncRequest: ", result)
 
         return_result = {}
         if state == "Completed":
@@ -946,13 +944,26 @@ class ToolingApi():
             }
 
         while state == "Queued":
-            util.append_message(panel, "ContainerAsyncRequest is in Queued, Waiting...")
+            Printer.get('log').write("ContainerAsyncRequest is in Queued, Waiting...")
             time.sleep(2)
 
             result = self.get(sync_request_url + "/" + request_id)
             state = result["State"]
             if state == "Completed":
                 return_result = {"success": True}
+
+        # Fix issue #74
+        if state in ["Error", "Aborted", "Invalidated"]:
+            if "ErrorMsg" in result:
+                problem = result["ErrorMsg"]
+            else:
+                problem = "This job is " + state
+            return_result = {
+                "success": False,
+                "problem": problem,
+                "columnNumber": 0,
+                "lineNumber": 0
+            }
 
         if state == "Failed":
             # This error need more process, because of confused single quote
@@ -966,12 +977,25 @@ class ToolingApi():
             if len(compile_errors) > 0:
                 return_result = compile_errors[0]
             else:
-                return_result["Error Message"] = result["ErrorMsg"]
+                return_result["problem"] = result["ErrorMsg"]
+                return_result["columnNumber"] = 0
+                return_result["lineNumber"] = 0
+
+            if "line" in return_result:
+                return_result["lineNumber"] = return_result["line"]
+                del return_result["line"]
+
+            if "column" in return_result:
+                return_result["columnNumber"] = return_result["column"]
+                del return_result["column"]
+
+            if "columnNumber" not in return_result:
+                return_result["columnNumber"] = 0
 
             return_result["success"] =  False
         
         if return_result["success"] and component_type == "ApexClass":
-            util.append_message(panel, "Start to fetch symbol table")
+            Printer.get('log').write("Start to fetch symbol table")
             query = "SELECT Id, SymbolTable " +\
                     "FROM ApexClassMember WHERE Id ='%s'" % member_result["id"]
             member = self.query(query, True)